rename binary to entangled
