+++ .github/workflows/test.yml
@@ -110,7 +110,7 @@ jobs:
       - uses: cachix/install-nix-action@v18
       - uses: cachix/cachix-action@v12
         with:
-          name: ethermint
+          name: entangle
           signingKey: "${{ secrets.CACHIX_SIGNING_KEY }}"
       - uses: technote-space/get-diff-action@v6.1.1
         with:
@@ -145,7 +145,7 @@ jobs:
       - uses: cachix/install-nix-action@v18
       - uses: cachix/cachix-action@v12
         with:
-          name: ethermint
+          name: entangle
           signingKey: "${{ secrets.CACHIX_SIGNING_KEY }}"
       - name: 'instantiate integration test env'
         run: nix-store -r $(nix-instantiate tests/integration_tests/shell.nix)
