+++ app/app.go
@@ -122,10 +122,10 @@ func init() {
 		panic(err)
 	}
 
-	DefaultNodeHome = filepath.Join(userHomeDir, ".ethermintd")
+	DefaultNodeHome = filepath.Join(userHomeDir, ".entangled")
 }
 
-const appName = "ethermintd"
+const appName = "entangled"
 
 var (
 	// DefaultNodeHome default home directories for the application daemon
