+++ contrib/scripts/test_localnet_liveness.sh
@@ -30,7 +30,7 @@ if [ -z "$4" ]; then
   exit 1
 fi
 
-docker_containers=($(docker ps -q -f name=ethermintd --format='{{.Names}}'))
+docker_containers=($(docker ps -q -f name=entangled --format='{{.Names}}'))
 
 while [ ${CNT} -lt $ITER ]; do
   curr_block=$(curl -s $NODEADDR:26657/status | jq -r '.result.sync_info.latest_block_height')
