+++ scripts/start-docker.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
 echo "prepare genesis: Run validate-genesis to ensure everything worked and that the genesis file is setup correctly"
-./ethermintd validate-genesis --home /ethermint
+./entangled validate-genesis --home /ethermint
 
 echo "starting ethermint node $ID in background ..."
-./ethermintd start \
+./entangled start \
 --home /ethermint \
 --keyring-backend test
