+++ tests/integration_tests/configs/default.jsonnet
@@ -1,7 +1,7 @@
 {
   dotenv: '../../../scripts/.env',
   'ethermint_9000-1': {
-    cmd: 'ethermintd',
+    cmd: 'entangled',
     'start-flags': '--trace',
     config: {
       consensus: {
