+++ tests/integration_tests/network.py
@@ -11,7 +11,7 @@ from web3.middleware import geth_poa_middleware
 from .cosmoscli import CosmosCLI
 from .utils import wait_for_port
 
-DEFAULT_CHAIN_BINARY = "ethermintd"
+DEFAULT_CHAIN_BINARY = "entangled"
 
 
 class Ethermint:
