+++ tests/integration_tests/test_upgrade.py
@@ -55,7 +55,7 @@ def post_init(path, base_port, config):
                 {
                     "command": f"cosmovisor start --home %(here)s/node{i}",
                     "environment": (
-                        f"DAEMON_NAME=ethermintd,DAEMON_HOME=%(here)s/node{i}"
+                        f"DAEMON_NAME=entangled,DAEMON_HOME=%(here)s/node{i}"
                     ),
                 }
             )
@@ -80,7 +80,7 @@ def custom_ethermint(tmp_path_factory):
         26100,
         Path(__file__).parent / "configs/cosmovisor.jsonnet",
         post_init=post_init,
-        chain_binary=str(path / "upgrades/genesis/bin/ethermintd"),
+        chain_binary=str(path / "upgrades/genesis/bin/entangled"),
     )
 
 
@@ -136,7 +136,7 @@ def test_cosmovisor_upgrade(custom_ethermint: Ethermint):
     # update cli chain binary
     custom_ethermint.chain_binary = (
         Path(custom_ethermint.chain_binary).parent.parent.parent
-        / f"{plan_name}/bin/ethermintd"
+        / f"{plan_name}/bin/entangled"
     )
     cli = custom_ethermint.cosmos_cli()
