+++ tests/rpc/rpc_pending_test.go
@@ -1,5 +1,5 @@
 // This is a test utility for Ethermint's Web3 JSON-RPC services.
-// To run these tests please first ensure you have the ethermintd running
+// To run these tests please first ensure you have the entangled running
 // You can configure the desired HOST and MODE as well in integration-test-all.sh
 package rpc
