+++ tests/solidity/test-helper.js
@@ -26,7 +26,7 @@ function checkTestEnv() {
     .describe('network', 'set which network to use: ganache|ethermint')
     .describe('batch', 'set the test batch in parallelized testing. Format: %d-%d')
     .describe('allowTests', 'only run specified tests. Separated by comma.')
-    .boolean('verbose-log').describe('verbose-log', 'print ethermintd output, default false')
+    .boolean('verbose-log').describe('verbose-log', 'print entangled output, default false')
     .argv;
 
   if (!fs.existsSync(path.join(__dirname, './node_modules'))) {
@@ -176,29 +176,29 @@ function setupNetwork({ runConfig, timeout }) {
   // Spawn the ethermint process
 
   const spawnPromise = new Promise((resolve, reject) => {
-    const ethermintdProc = spawn('./init-test-node.sh', {
+    const entangledProc = spawn('./init-test-node.sh', {
       cwd: __dirname,
       stdio: ['ignore', runConfig.verboseLog ? 'pipe' : 'ignore', 'pipe'],
     });
 
-    logger.info(`Starting Ethermintd process... timeout: ${timeout}ms`);
+    logger.info(`Starting entangled process... timeout: ${timeout}ms`);
     if (runConfig.verboseLog) {
-      ethermintdProc.stdout.pipe(process.stdout);
+      entangledProc.stdout.pipe(process.stdout);
     }
-    ethermintdProc.stderr.on('data', d => {
+    entangledProc.stderr.on('data', d => {
       const oLine = d.toString();
       if (runConfig.verboseLog) {
         process.stdout.write(oLine);
       }
       if (oLine.indexOf('Starting JSON-RPC server') !== -1) {
-        logger.info('Ethermintd started');
-        resolve(ethermintdProc);
+        logger.info('entangled started');
+        resolve(entangledProc);
       }
     });
   });
 
   const timeoutPromise = new Promise((resolve, reject) => {
-    setTimeout(() => reject(new Error('Start ethermintd timeout!')), timeout);
+    setTimeout(() => reject(new Error('Start entangled timeout!')), timeout);
   });
   return Promise.race([spawnPromise, timeoutPromise]);
 }
